requirements.txt
aioredis==1.3.1
asgiref==3.3.4
astroid==2.5.6
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
backcall==0.2.0
certifi==2021.5.30
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
decorator==5.0.9
dj-database-url==0.5.0
Django==3.2.4
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
ipython==7.24.1
ipython-genutils==0.2.0
isort==5.8.0
jedi==0.18.0
lazy-object-proxy==1.6.0
matplotlib-inline==0.1.2
mccabe==0.6.1
msgpack==1.0.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pylint==2.8.3
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyOpenSSL==20.0.1
pytz==2021.1
requests==2.25.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
Twisted==21.2.0
txaio==21.2.1
urllib3==1.26.5
wcwidth==0.2.5
whitenoise==5.2.0
wrapt==1.12.1
zope.interface==5.4.0
